rewrite to be a reader/writer/cleaner
