+++ ifucube/__init__.py
@@ -1,0 +1,3 @@
+from .wavelength import *
+from .ifucube import *
+from .ifucubelist import *
